feat: main-view formatDateRange 함수 Extension으로 공통화
