+++ C3-4T2D/Extensions/DateExtensions.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+extension DateFormatter {
+    static func projectDateRange(startDate: Date, endDate: Date?) -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy.MM.dd"
+
+        if let endDate = endDate {
+            return "\(formatter.string(from: startDate)) ~ \(formatter.string(from: endDate))"
+        } else {
+            return "\(formatter.string(from: startDate)) ~"
+        }
+    }
+}
