+++ C3-4T2D/View/MainView/MainView.swift
@@ -23,7 +23,7 @@ struct MainView: View {
                             Text("진행중인 과정").font(.system(size: 22, weight: .bold)).padding(.bottom, 8)
                             Text(dummyData.currentProject.projectTitle).font(.system(size: 19, weight: .bold))
                             HStack {
-                                Text(formatDateRange(
+                                Text(DateFormatter.projectDateRange(
                                     startDate: dummyData.currentProject.postList.compactMap { $0.createdAt }.min() ?? Date(),
                                     endDate: dummyData.currentProject.finishedAt
                                 ))
@@ -124,139 +124,6 @@ struct MainView: View {
                 .padding(.vertical, 20)
             }
         }
-    }
-}
-
-private func formatDateRange(startDate: Date, endDate: Date?) -> String {
-    let formatter = DateFormatter()
-    formatter.dateFormat = "yyyy.MM.dd"
-
-    if let endDate = endDate {
-        return "\(formatter.string(from: startDate)) ~ \(formatter.string(from: endDate))"
-    } else {
-        return "\(formatter.string(from: startDate)) ~"
-    }
-}
-
-// 프로젝트 스크롤뷰
-struct ProjectSectionCard: View {
-    let project: Project
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(project.projectTitle)
-                        .font(.system(size: 19, weight: .bold))
-
-                    HStack {
-                        Text(formatDateRange(
-                            startDate: project.postList.compactMap { $0.createdAt }.min() ?? Date(),
-                            endDate: project.finishedAt
-                        ))
-                        .font(.system(size: 11, weight: .regular))
-                        .foregroundColor(.gray)
-
-                        Text(project.finishedAt == nil ? "진행중" : "완료")
-                            .font(.system(size: 11, weight: .medium))
-                            .padding(.horizontal, 8)
-                            .padding(.vertical, 2)
-                            .background(project.finishedAt == nil ? Color.green.opacity(0.2) : Color.gray.opacity(0.2))
-                            .foregroundColor(project.finishedAt == nil ? .green : .gray)
-                            .cornerRadius(4)
-                    }
-                }
-
-                Spacer()
-
-                Image(systemName: "chevron.right")
-                    .foregroundColor(.gray)
-                    .font(.system(size: 20, weight: .bold))
-            }.padding(.trailing, 20)
-
-            ScrollView(.horizontal, showsIndicators: false) {
-                HStack(spacing: 12) {
-                    ForEach(project.postList) { _ in
-                        Image("tmpImage")
-                            .resizable()
-                            .frame(width: 140, height: 100)
-                            .cornerRadius(8)
-                    }
-                    if project.postList.count > 4 {
-                        ZStack {
-                            Color.gray.opacity(0.1)
-                                .frame(width: 80, height: 80)
-                                .cornerRadius(8)
-
-                            Text("+\(project.postList.count - 4)")
-                                .font(.system(size: 14, weight: .medium))
-                                .foregroundColor(.gray)
-                        }
-                    }
-                }
-                .padding(.trailing, 20)
-            }
-        }
-    }
-}
-
-// 리스트뷰
-struct PostListCard: View {
-    let project: Project
-
-    var body: some View {
-        HStack(spacing: 15) {
-            Image("tmpImage")
-                .resizable()
-                .frame(width: 100, height: 70)
-                .cornerRadius(5)
-
-            VStack(alignment: .leading, spacing: 5) {
-                Text(project.projectTitle)
-                    .font(.system(size: 17, weight: .semibold))
-                    .lineLimit(2)
-
-                HStack {
-                    Text(formatDateRange(
-                        startDate: project.postList.compactMap { $0.createdAt }.min() ?? Date(),
-                        endDate: project.finishedAt
-                    ))
-                    .font(.system(size: 11, weight: .regular))
-                    .foregroundColor(.gray)
-
-                    Text(project.finishedAt == nil ? "진행중" : "완료")
-                        .font(.system(size: 10, weight: .medium))
-                        .padding(.horizontal, 5)
-                        .padding(.vertical, 2)
-                        .background(project.finishedAt == nil ? Color.green.opacity(0.2) : Color.gray.opacity(0.2))
-                        .foregroundColor(project.finishedAt == nil ? .green : .gray)
-                        .cornerRadius(4)
-                }
-
-                HStack(spacing: 12) {
-                    HStack(spacing: 4) {
-                        Image(systemName: "folder.fill")
-                            .foregroundColor(.orange)
-                            .font(.system(size: 12))
-                        Text("12")
-                            .font(.system(size: 12, weight: .semibold))
-                    }
-
-                    HStack(spacing: 4) {
-                        Image(systemName: "person.fill")
-                            .foregroundColor(.orange)
-                            .font(.system(size: 12, weight: .semibold))
-                        Text("112")
-                            .font(.system(size: 14, weight: .medium))
-                    }
-                }
-
-                Spacer()
-            }
-
-            Spacer()
-        }
-        Divider()
     }
 }
